backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
gunicorn==21.2.0
supabase==2.3.4
httpx==0.26.0